Create and refactor cli to module
