+++ nodes/node/binary/src/main.rs
@@ -2,10 +2,8 @@ use clap::Parser as _;
 use color_eyre::eyre::{Result, eyre};
 use logos_blockchain_node::{
     UserConfig,
-    config::{
-        CliArgs, DeploymentType, OnUnknownKeys, deployment::DeploymentSettings,
-        deserialize_config_at_path,
-    },
+    cli::{CliArgs, Command, build_run_config},
+    config::{DeploymentType, OnUnknownKeys, deployment::DeploymentSettings, deserialize_config_at_path},
     get_services_to_start, run_node_from_config,
 };
 
@@ -18,18 +16,18 @@ async fn main() -> Result<()> {
 
     if let Some(command) = cli_args.command {
         match command {
-            logos_blockchain_node::config::Command::Init(init_args) => {
-                return logos_blockchain_node::init::run(&init_args);
+            Command::Init(init_args) => {
+                return logos_blockchain_node::cli::init::run(&init_args);
             }
-            logos_blockchain_node::config::Command::Inscribe(inscribe_args) => {
+            Command::Inscribe(inscribe_args) => {
                 lb_tui_zone::run(inscribe_args).await;
                 return Ok(());
             }
-            logos_blockchain_node::config::Command::Participate(participate_args) => {
-                return logos_blockchain_node::participate::run(&participate_args);
+            Command::Participate(participate_args) => {
+                return logos_blockchain_node::cli::participate::run(&participate_args);
             }
-            logos_blockchain_node::config::Command::GetPeerId(get_peer_id_args) => {
-                return logos_blockchain_node::get_peer_id::run(&get_peer_id_args);
+            Command::GetPeerId(get_peer_id_args) => {
+                return logos_blockchain_node::cli::get_peer_id::run(&get_peer_id_args);
             }
         }
     }
@@ -68,7 +66,7 @@ async fn main() -> Result<()> {
                 .inspect_err(|e| {
                 eprintln!("\nExiting... {e}.\n");
             })?;
-        user_config.update_from_args(cli_args)?
+        build_run_config(user_config, cli_args)?
     };
 
     let app = run_node_from_config(run_config, None)
